Fix: Move app route & controller to api folder root
